fct respond to feliz jueves link one user only
TO DO: implement a json file or text file or db to store users infos (name, count)
